.human/ansible/inventory_setup/requirements.txt
PySide6
pyyaml